+++ Save/Extensions/UIApplicationExtension.swift
@@ -1,0 +1,15 @@
+//
+//  UIApplicationExtension.swift
+//  Save
+//
+//  Created by navoda on 2025-06-10.
+//  Copyright © 2025 Open Archive. All rights reserved.
+//
+
+import UIKit
+
+extension UIApplication {
+    func endEditing() {
+        sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+    }
+}
