+++ Save/Settings/PrivateServer/ServerSettingsView.swift
@@ -181,142 +181,3 @@ struct ServerSettingsView: View {
         
     }}
 
-   // MARK: - Custom TextField
-
-
-struct CustomTextField: View {
-    var placeholder: String
-    @Binding var text: String
-    var isSecure: Bool = false
-    var isDisabled: Bool = false
-    var onEditingChanged: (() -> Void)? = nil
-    var onCommit: (() -> Void)? = nil
-   
-
-    var body: some View {
-        ZStack(alignment: .leading) {
-            if text.isEmpty {
-                Text(placeholder)
-                    .italic()
-                    .font(.montserrat(.medium, for: .footnote))
-                    .foregroundColor(.textEmpty)
-                    .padding(.leading, 16)
-            }
-
-            if isSecure {
-                SecureField("", text: $text)
-                    .font(.montserrat(.medium, for: .footnote))
-                    .padding(12)
-            } else {
-                if #available(iOS 14.0, *) {
-                    TextField("", text: $text, onEditingChanged: { _ in
-                        onEditingChanged?()
-                    }, onCommit: {
-                        onCommit?()
-                    })
-                    .onChange(of: text) { _ in
-                        onEditingChanged?()
-                    }
-                    .font(.montserrat(.medium, for: .footnote))
-                    .padding(12)
-                } else {
-                    // Add fallback logic if needed
-                    TextField("", text: $text)
-                        .font(.montserrat(.medium, for: .footnote))
-                        .padding(12)
-                }
-            }
-        }
-        .frame(height: 50)
-        .cornerRadius(8)
-        .overlay(RoundedRectangle(cornerRadius: 8).stroke(Color.gray.opacity(0.7)))
-        .background(isDisabled ? Color.gray.opacity(0.2) : Color.textboxBg)
-        .disabled(isDisabled)
-        .padding(.bottom, 8)
-    }
-
-}
-
-   // MARK: - License Toggles
-   struct LicenseToggles: View {
-       @ObservedObject var store: ServerSettingsStore
-
-       var body: some View {
-           VStack(alignment: .leading, spacing: 10) {
-               Toggle(NSLocalizedString("Allow anyone to remix and share?", comment: "Remix Toggle"), isOn: Binding(
-                   get: { store.state.allowRemix },
-                   set: { newValue in
-                       store.dispatch(action: .toggleAllowRemix(newValue))
-                       store.dispatch(action: .updateLicense)
-                   }
-               )) .toggleTint(.accent).font(.montserrat(.medium, for: .subheadline))
-
-               Toggle(NSLocalizedString("Require them to share like you have?", comment: "ShareAlike Toggle"), isOn: Binding(
-                   get: { store.state.requireShareAlike },
-                   set: { newValue in
-                       store.dispatch(action: .toggleRequireShareAlike(newValue))
-                       store.dispatch(action: .updateLicense)
-                   }
-               )) .toggleTint(.accent)
-               .disabled(!store.state.allowRemix).font(.montserrat(.medium, for: .subheadline))
-
-               Toggle(NSLocalizedString("Allow commercial use?", comment: "Commercial Use Toggle"), isOn: Binding(
-                   get: { store.state.allowCommercialUse },
-                   set: { newValue in
-                       store.dispatch(action: .toggleAllowCommercialUse(newValue))
-                       store.dispatch(action: .updateLicense)
-                   }
-               )) .toggleTint(.accent).font(.montserrat(.medium, for: .subheadline))
-         
-               if let licenseURL = store.state.licenseURL {
-                   if #available(iOS 14.0, *) {
-                       Link(NSLocalizedString(licenseURL, comment: "License Link"), destination: URL(string: licenseURL)!)
-                           .foregroundColor(.accentColor)
-                           .padding(.top, 10) .font(.montserrat(.medium, for: .subheadline))
-                   } else {
-                       // Fallback on earlier versions
-                   }
-               }
-
-           }
-       }
-   }
-
-   // MARK: - Section Header
-   struct SectionHeader: View {
-       var title: String
-
-       var body: some View {
-           Text(NSLocalizedString(title, comment: "\(title) Section"))
-               .font(.montserrat(.semibold, for: .headline))
-               .foregroundColor(.gray70)
-               .padding(.top,20)
-               .padding(.bottom,10)
-       }
-   
-}
-
-
-struct ToggleTintModifier: ViewModifier {
-    var color: Color
-
-    func body(content: Content) -> some View {
-        if #available(iOS 15.0, *) {
-            content.tint(color) 
-        } else {
-            content.accentColor(color)
-        }
-    }
-}
-
-// MARK: - View Extension for Easy Usage
-extension View {
-    func toggleTint(_ color: Color) -> some View {
-        self.modifier(ToggleTintModifier(color: color))
-    }
-}
-extension UIApplication {
-    func endEditing() {
-        sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-    }
-}
